+++ changelog.md
@@ -1,3 +1,8 @@
 # Changelog
 
-Common test function split is now complete.
+**./app/paths/files/fox-paths.js**
+* 'foxRoot' in 'defineFoxPaths'
+	* Now uses a set path to the adjacent folder.
+	* This is how it was originally before emulation.
+* Removed the 'getFoxRoot' function.
+* Removed 'useEmulator' variable from 'defineFoxPaths'